requirements.txt
streamlit
google-generativeai
requests
pdf2image
pillow
pandas
openpyxl
